+++ appdaemon/utils.py
@@ -6,6 +6,51 @@ import random
 import uuid
 import asyncio
 
+class Formatter(object):
+    def __init__(self):
+        self.types = {}
+        self.htchar = '\t'
+        self.lfchar = '\n'
+        self.indent = 0
+        self.set_formater(object, self.__class__.format_object)
+        self.set_formater(dict, self.__class__.format_dict)
+        self.set_formater(list, self.__class__.format_list)
+        self.set_formater(tuple, self.__class__.format_tuple)
+
+    def set_formater(self, obj, callback):
+        self.types[obj] = callback
+
+    def __call__(self, value, **args):
+        for key in args:
+            setattr(self, key, args[key])
+        formater = self.types[type(value) if type(value) in self.types else object]
+        return formater(self, value, self.indent)
+
+    def format_object(self, value, indent):
+        return repr(value)
+
+    def format_dict(self, value, indent):
+        items = [
+            self.lfchar + self.htchar * (indent + 1) + repr(key) + ': ' +
+            (self.types[type(value[key]) if type(value[key]) in self.types else object])(self, value[key], indent + 1)
+            for key in value
+        ]
+        return '{%s}' % (','.join(items) + self.lfchar + self.htchar * indent)
+
+    def format_list(self, value, indent):
+        items = [
+            self.lfchar + self.htchar * (indent + 1) + (self.types[type(item) if type(item) in self.types else object])(self, item, indent + 1)
+            for item in value
+        ]
+        return '[%s]' % (','.join(items) + self.lfchar + self.htchar * indent)
+
+    def format_tuple(self, value, indent):
+        items = [
+            self.lfchar + self.htchar * (indent + 1) + (self.types[type(item) if type(item) in self.types else object])(self, item, indent + 1)
+            for item in value
+        ]
+        return '(%s)' % (','.join(items) + self.lfchar + self.htchar * indent)
+
 
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 requests.packages.urllib3.disable_warnings(InsecureRequestWarning)
@@ -14,6 +59,46 @@ constraints = (
     "constrain_input_select", "constrain_presence",
     "constrain_start_time", "constrain_end_time"
 )
+
+
+class AttrDict(dict):
+    """ Dictionary subclass whose entries can be accessed by attributes
+        (as well as normally).
+    """
+    def __init__(self, *args, **kwargs):
+        super(AttrDict, self).__init__(*args, **kwargs)
+        self.__dict__ = self
+
+    @staticmethod
+    def from_nested_dict(data):
+        """ Construct nested AttrDicts from nested dictionaries. """
+        if not isinstance(data, dict):
+            return data
+        else:
+            return AttrDict({key: AttrDict.from_nested_dict(data[key])
+                                for key in data})
+
+
+class StateAttrs(dict):
+
+    def __init__(self, dict):
+        device_dict = {}
+        devices = set()
+        for entity in dict:
+            if "." in entity:
+                device, name = entity.split(".")
+                devices.add(device)
+        for device in devices:
+            entity_dict = {}
+            for entity in dict:
+                if "." in entity:
+                    thisdevice, name = entity.split(".")
+                    if device == thisdevice:
+                        entity_dict[name] = dict[entity]
+            device_dict[device] = AttrDict.from_nested_dict(entity_dict)
+
+        self.__dict__ = device_dict
+
 
 
 def _secret_yaml(loader, node):